Renomenando arquivos de execução dos casos de teste e excluindo codigos repetidos
