+++ IC/Comando de repetições/Veiculos trafegando em alta velocidade/Testes/test.py
@@ -1,9 +1,10 @@
 import unittest
 import sys
 import io
+import os
 
 
-def verificar_string(string1, valores_entrada,arquivo):
+def verificar_string(valores_esperados, valores_entrada,arquivo):
     # Abre o arquivo 'codigo.py' e lê o seu conteúdo
     with open(arquivo, 'r') as file:
         codigo = file.read()
@@ -22,7 +23,11 @@ def verificar_string(string1, valores_entrada,arquivo):
     try:
         # Executa o código lido do arquivo com o input mockado
         exec(codigo, {'input': input_mock_function})
-        valor_impresso1 = sys.stdout.getvalue().strip()
+        valor_impresso = sys.stdout.getvalue().strip()
+        # partes = valor_impresso.split('\n')
+        # saida1 = partes[0]
+        # saida2 = partes[1]
+        # saida3 = partes[2]
 
     except Exception as e:
         print(f"Erro ao executar o código: {e}")
@@ -32,38 +37,48 @@ def verificar_string(string1, valores_entrada,arquivo):
         sys.stdout = stdout_backup
 
     # Verifica se os valores impressos são iguais às strings fornecidas
-    return string1 == valor_impresso1
+    if(valores_esperados == valor_impresso):
+        return True
+    else:
+        return False
 
 
 class TestStringVerification(unittest.TestCase):
     #file = 'codigo.py'
     def test_verificar_string_1(self):
-        valores_entrada = ["01092023"]
-        self.assertTrue(verificar_string("01 de setembro de 2023", valores_entrada,self.file))
-
+        valores_entrada = ["[60.0, 73.9, 22.9, 72.1, 72.0, 100.6]"] 
+        self.assertTrue(verificar_string("1\n3\n2", valores_entrada,self.file))
+        
     def test_verificar_string_2(self):
-        valores_entrada = ["17031909"]
-        self.assertTrue(verificar_string("17 de marco de 1909", valores_entrada,self.file))
+        valores_entrada = ["[11.6, 60.6, 16.6, 11.35]"]
+        self.assertTrue(verificar_string("2\n1", valores_entrada,self.file))
 
     def test_verificar_string_3(self):
-        valores_entrada = ["31021031"]
-        self.assertTrue(verificar_string("31 de fevereiro de 1031", valores_entrada,self.file))
+        valores_entrada = ["[72.9, 29.1, 7.29, 90.7, 66.6]"]
+        self.assertTrue(verificar_string("3\n1", valores_entrada,self.file))
 
-def runTest(arg1, arg2, arg3, arg4):
+def runTest(nameLLm, prompt, language, outDir):
     import xmlrunner as r
-    import glob
-
-    file = glob.glob(f"/home/**/Repositorio_IC/**/Data*/**/{arg4}/{arg1}{arg2}{arg3}.py", recursive=True)
-    #print(file)
+    import glob 
+       
+    file = glob.glob(f"{outDir}/**/{nameLLm}{prompt}{language}.py", recursive=True)
     TestStringVerification.file = file[0]
-    #unittest.main(argv=sys.argv[:1], testRunner=r.XMLTestRunner(output='testeData', outsuffix="resultado Data"))
     suite = unittest.TestLoader().loadTestsFromTestCase(TestStringVerification)
-    runner = r.XMLTestRunner(output='testeData', outsuffix="resultado_Data")
+    outDir = outDir +"/XML"
+    if not(os.path.exists(outDir)):
+        os.makedirs(outDir,exist_ok=True)
+    runner = r.XMLTestRunner(output=outDir, outsuffix=f"{prompt}resultado_Veiculos")
     runner.run(suite)
+
+    del(glob)
+    del(r)
+    #outsuffix = time.strftime("%Y%m%d%H%M%S")
 
 if __name__ == '__main__':
     #parametros que devem ser passados:
     #arg1 -> nome LLM
     #arg2 -> qual parte do prompt
-    #arg3 ->idioma (pt, eng, eng2)  
+    #arg3 ->idioma (pt, eng, eng2)
     runTest(arg1=sys.argv[1], arg2=sys.argv[2], arg3=sys.argv[3])
+   
+
