--- codeEvaluator.py
@@ -1,16 +1,0 @@
-import glob
-import generativeIA as 
-class CodeEvaluator:
-    def __init__(self, output_dirctory, input_dirctory, language, prompt, qtd)
-        self.base_output_dirctory = output_dirctory
-        self.base_input_dirctory = input_dirctory
-        self.language = language
-        self.promppt_type = prompt
-        self.k = qtd #quantas vezes ele vai gerar
-
-    def run(self,)
-    def discolver_problems(self,):
-        list problems = glob.glob('/**/Repositorio_IC/**/Prompts/*.txt',recursive=true)
-        return problems
-    def createLLM(self,) -> None:
-        pass
